Changes to Abstract the AzureService so allowing different implementations to be injected - AzureServiceCloud or AzureServiceStack dependant on the value of azure.variant property (cloud|stack)
