+++ analysis-platform-controller/src/main/resources/application.properties
@@ -1,4 +1,5 @@
 # Azure service properties
+azure.variant=cloud
 azure.activedirectory.endpoint=XXXX
 azure.login.clientid=XXXX
 azure.login.domain=XXXX
